refactor: 💡 getOrCreateValue를 Map과 Object 둘 다 사용 가능하도록 오버로딩
