+++ backend/src/interests/repository/interests.in-memory.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { InterestsManager } from '../operator/Interests.manager';
-import { getOrCreateValue } from '../../common/util/map.function';
+import { getOrCreateValue } from '../../common/util/get-or-create-value';
 import { Interest } from '../domain/interest';
 import * as AsyncLock from 'async-lock';
 import { CustomException } from '../../common/exception/custom-exception';
